pyproject.toml
[project]
name = "rosalind-problems"
version = "0.1.0"
description = "Repo for code related to to solving rosalind problem and the solutions"
readme = "README.md"
authors = [
  {name = "Aminata Tengbe", email = "a.l.tengbe@proton.me"}
]
maintainers = [
  {name = "Aminata Tengbe", email = "a.l.tengbe@proton.me"}
]
classifiers = [
    # TODO
]
license = {text = "MIT"}
dependencies = [
  "typer",
  "beautifulsoup4 (>=4.13.4,<5.0.0)",
  "requests (>=2.32.4,<3.0.0)",
  "html-to-markdown (>=1.9.0,<2.0.0)",
  "gitpython (>=3.1.45,<4.0.0)",
]
requires-python = ">= 3.10"

[project.optional-dependencies]
test = [
    "coverage",  # testing
    "pytest",  # testing
    "ruff",  # linting
    "ty", # checking types
    "ipdb", # debugging
]

[project.urls]
bugs = "https://github.com/natleahh/rosalind_problems/issues"
changelog = "https://github.com/natleahh/rosalind_problems/blob/master/changelog.md"
homepage = "https://github.com/natleahh/rosalind_problems"

[project.scripts]
get_problem = "rosalind_problems.scripts.get_problem:get_problem"

[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.uv]
package = true

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report term-missing"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
requests-mock = "^1.12.1"

